fix(game): raise the evidence bar of the automatic frame rate limit
Review findings on the draft. A descent on the two-second fast rule is
now always provisional and owes its confirming probe, only a settled
verdict is stored, and the frames that follow a loading cover are no
reading at all: one stutter episode on a capable machine could pin it at
half rate for ten minutes and into its next session. A hidden web tab, a
suspend or a long stall is no play time (the resume interval was credited
whole). The arm flag is set after the arm call returns, a persistent
failure logs once, the singleton is assigned last, junk in ?fpscap is no
override. An automatic limit above the old 45 fps line reads as a good
session for the perf doctor. An abandoned cap probe that lost its origin
restores the band baselines. refresh_hz is an INT. Docs and tests follow.

+++ server/client_perf_reports_schema.ts
@@ -125,10 +125,11 @@ ALTER TABLE client_perf_reports ADD COLUMN IF NOT EXISTS desktop_shell BOOLEAN N
 -- raw_summary.cadence block that also says so is shed under a size squeeze.
 -- frame_cap_intent is 0 (none), 30 or 60; cadence_divisor is how many display
 -- refresh slots one rendered frame spans (1 = ceiling inert); refresh_hz is
--- the client's estimate of the display rate (0 = unknown, which with an intent
+-- the client's estimate of the display rate in WHOLE Hz (an INT on purpose:
+-- a finer reading would fingerprint the display) (0 = unknown, which with an intent
 -- set is the unpaced limiter). target_fps is the effective target from the
 -- same release on. Pre-column rows read as no ceiling, which is what they had.
 ALTER TABLE client_perf_reports ADD COLUMN IF NOT EXISTS frame_cap_intent INT NOT NULL DEFAULT 0;
 ALTER TABLE client_perf_reports ADD COLUMN IF NOT EXISTS cadence_divisor INT NOT NULL DEFAULT 1;
-ALTER TABLE client_perf_reports ADD COLUMN IF NOT EXISTS refresh_hz REAL NOT NULL DEFAULT 0;
+ALTER TABLE client_perf_reports ADD COLUMN IF NOT EXISTS refresh_hz INT NOT NULL DEFAULT 0;
 `;
